# tools/dashboard/backend/docs/swagger.yaml
definitions:
...
  service.DiagnosticJob:
    properties:
      cluster:
        type: string
      completedAt:
        type: string
      id:
        type: string
      message:
        type: string
      namespace:
        type: string
      outputPath:
        type: string
      progress:
        type: integer
      startedAt:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/v1/diagnostics/{namespace}/{cluster}/start:
    post:
      consumes:
      - application/json
      description: Create a polardbx-clinic Pod to collect diagnostic information
        for the given cluster.
      parameters:
      - description: Kubernetes namespace of the cluster
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the PolarDB-X cluster
        in: path
        name: cluster
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Diagnostic task started
          schema:
            $ref: '#/definitions/service.DiagnosticJob'
        "400":
          description: Request parameter error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Start diagnosis
      tags:
      - diagnostics
  /api/v1/diagnostics/{namespace}/{id}:
    delete:
      consumes:
      - application/json
      description: Delete diagnostic Pod and related resources
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: diagnostic task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deletion successful
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete diagnostic task
      tags:
      - diagnostics
  /api/v1/diagnostics/{namespace}/{id}/download:
    get:
      consumes:
      - application/json
      description: Get download link or path for diagnostic report
      parameters:
      - description: Kubernetes namespace of the diagnostic job
        in: path
        name: namespace
        required: true
        type: string
      - description: diagnostic task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Download information including path, URL, and helper command
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request parameter error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Diagnostic report not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Download diagnostic report
      tags:
      - diagnostics
  /api/v1/diagnostics/{namespace}/{id}/file:
    get:
      description: Get diagnostic report file content from Pod
      parameters:
      - description: Kubernetes namespace of the diagnostic job
        in: path
        name: namespace
        required: true
        type: string
      - description: diagnostic task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/gzip
      responses:
        "200":
          description: Diagnostic report file
          schema:
            type: file
        "400":
          description: Request parameter error (e.g., invalid ID)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required or invalid authentication state
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Diagnostic job or file not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Diagnostic report not ready yet
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get diagnostic report file
      tags:
      - diagnostics
  /api/v1/diagnostics/{namespace}/{id}/status:
    get:
      consumes:
      - application/json
      description: Get current status and progress of specified diagnostic task
      parameters:
      - description: Kubernetes namespace of the diagnostic job
        in: path
        name: namespace
        required: true
        type: string
      - description: diagnostic task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Diagnostic task status
          schema:
            $ref: '#/definitions/service.DiagnosticJob'
        "400":
          description: Request parameter error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Diagnostic task not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get diagnostic status
      tags:
      - diagnostics
  /api/v1/diagnostics/reports:
    get:
      consumes:
      - application/json
      description: List all diagnostic reports under specified namespace
      parameters:
      - description: Kubernetes namespace; list all namespaces if not specified
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Diagnostic reports list
          schema:
            items:
              $ref: '#/definitions/service.DiagnosticJob'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List diagnostic reports
      tags:
      - diagnostics
...
swagger: "2.0"